Add more context/framing in an "Article 0"
Even before voting on the proposed constitution began, many ENS community members noticed that the text of the constitution does not describe the broader context of the ENS community, which existed before the ENS token and associated DAO, processes, etc. were introduced. However, this context/framing is considered to be very important in interpreting what the constitution document is or can/should be. So the idea emerged to add an "Article 0" (both in reference to the zeroth law from Asimov's robotics series and in reference to general computer science numbering conventions) which would both serve as a preamble/introduction and try to express the pre-existing and mostly not-mechanically-enforceable values of the ENS community within which the constitution was being proposed. With input and feedback from many community members, the text was mostly written by @emansipater with some edits and feedback from @Arachnid as well as multiple rounds of review on Twitter and Discord from many others.

Although Article 0 emerged before the token airdrop and associated voting processes, it was not included in the initial voting interface for $ENS airdrop recipients and has instead proceeded through the phases of the governance process, with the "temperature check" phase adding many positive/supportive comments and a few slight suggested edits. Now it is time to make a formal "Draft" proposal with the exact diff of proposed changes. Because the original constitution document did not have an Article 0 I have made some minimal changes to the wording at the top for flow, before adding in the proposed Article 0 at the head of the numerical list. There is a slight tweak to the final paragraph of Article 0, based on feedback from the "temperature check" thread. Further feedback or changes can be discussed in the official draft proposal thread on discourse, keeping in mind that there is now quite a lot of positive feedback accumulated for the phrasing so smaller changes/edits are preferred unless large issues are identified. But all feedback (in all venues) is welcome, since the specific goal of Article 0 is to capture surrounding context which might otherwise be missed by the mechanics of the new ENS DAO processes themselves.