Create ep2.x(formerly-ep15)-dissolution-of-the-community-working-group.md
EP15 was named EP14 in error under the old naming convention resulting in a conflict and renaming the same to EP 15.  

However, even with the correction to EP15 under the old naming convention the file was erroneously included under term-4 in github.  To illustrate the problem of the old naming convention further, this EP is not include at all on the Governance Documentation located at: https://docs.ens.domains/v/governance/governance-proposals

This PR both moves the file to the proper term-2 on GitHub and renames the same to ep2.x(formerly ep15).  Unfortunately there is no correct enumeration under the revised naming convention; therefore, this is the sole EP/file name proposed with an "x" to fit between 2.1 and 2.2.1.  For purposes of clarity the file has been named "ep2.x(formerly ep15)...".